Makefile changes to support stand-alone builds
The biggest change is that you must specify the location of a postgres
source repo that's at least had configure (with appropriate TCL options)
run. Unfortunately, make check does not work (just like it doesn't with
PGXS). If you try to make check, you'll get an error. There's a few
other editorial changes.

Almost all required Makefile changes are in stand-alone.mk. See the new
comment at the top of the Makefile for details.

Note that because this still points at a live Postgres repo it's
possible to accidentally end up with the built-in version of TCL! For
now, the best way to protect against that is to either change one of the
error messages that's tested for (search for 'even columns') and make
sure you get a failure. Or, just temporarily move the tcl directory from
$(top_builddir)/src/pl to somewhere else. Keep in mind that because only
installcheck is supported, the actual testing is being done on a live
install somewhere on your machine.